requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
decorator==4.4.1
dj-database-url==0.5.0
Django==3.0.7
django-debug-toolbar==2.2
django-debug-toolbar-template-profiler==2.0.0
django-extensions==2.2.8
django-redis-cache==2.1.0
easy-thumbnails==2.7
elasticsearch==7.5.1
gunicorn==20.0.4
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
kombu==4.6.7
Markdown==3.2.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
python-decouple==3.3
python-http-client==3.2.7
pytz==2019.3
redis==3.4.1
sendgrid==6.4.1
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.0
starkbank-ecdsa==1.0.0
traitlets==4.3.3
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
wrapt==1.12.0
zipp==3.0.0
